Move libpcap into internal/
